requirements.txt
django==3.0.5  # pyup: < 3.1  # https://www.djangoproject.com/
djangorestframework==3.11.0  # https://github.com/encode/django-rest-framework
django-environ==0.4.5

psycopg2==2.8.5 --no-binary psycopg2  # https://github.com/psycopg/psycopg2

# Testing
# ------------------------------------------------------------------------------
pytest==5.4.3  # https://github.com/pytest-dev/pytest
pytest-sugar==0.9.3  # https://github.com/Frozenball/pytest-sugar
pytest-django==3.9.0

# Code quality
# ------------------------------------------------------------------------------
flake8==3.8.2  # https://github.com/PyCQA/flake8

# Django
# ------------------------------------------------------------------------------
factory-boy==2.12.0  # https://github.com/FactoryBoy/factory_boy
